Fix Sitemap loader.
