+++ libs/superagent/app/datasource/types.py
@@ -10,6 +10,7 @@ VALID_UNSTRUCTURED_DATA_TYPES = [
     "NOTION",
     "URL",
     "YOUTUBE",
+    "SITEMAP",
 ]
 
 VALID_STRUCTURED_DATA_TYPES = ["AIRTABLE", "CSV", "STRIPE", "XLSX"]
