+++ libs/superagent/pyproject.toml
@@ -53,7 +53,6 @@ langchain-experimental = "^0.0.37"
 pydub = "^0.25.1"
 langsmith = "^0.0.63"
 langchain = "^0.0.334"
-nest-asyncio = "^1.5.8"
 
 
 [build-system]
